Split `Transformer` functionality
